fix: address all code review issues (HIGH/MEDIUM/LOW)
- H1: fix integer overflow in assertImage (int→size_t for pixel counts)
- H2: fix localtime_s portability (add localtime_r for POSIX)
- H3: remove misleading std::once_flag (single-threaded server)
- M1: reject tools/list and tools/call before initialization (-32002)
- M2: throw on duplicate tool name registration
- M3: reject unknown parameters in validateArgs
- M4: remove dead null checks after controller() (already throws)
- M5: replace hand-crafted JSON in snapshot.cpp with nlohmann::json
- M6: nest AssertResult details under "details" key
- M7: isolate stb_image IMPLEMENTATION into dedicated TU
- L1: check std::cout.fail() to detect broken output pipe
- L2: fix #include order in main.cpp
- L3: tighten findNewestCapture filename matching
- L4: make test stub controller() throw like real implementation
- L5: add batch edge case tests (non-object elements, all-notifications)
- L6: add ShaderEditState::clear() method

+++ src/core/types.h
@@ -403,8 +403,8 @@ struct PixelAssertResult {
 
 struct ImageCompareResult {
     bool pass = false;
-    int diffPixels = 0;
-    int totalPixels = 0;
+    size_t diffPixels = 0;
+    size_t totalPixels = 0;
     double diffRatio = 0.0;
     std::string diffOutputPath;
     std::string message;
